Move SC integration tests to new project.
