+++ src/Stratis.Bitcoin.Features.SmartContracts/ReflectionExecutor/Controllers/ContractExplorerController.cs
@@ -1,0 +1,10 @@
+﻿//using Microsoft.AspNetCore.Mvc;
+
+//namespace Stratis.Bitcoin.Features.SmartContracts.ReflectionExecutor.Controllers
+//{
+//    [Route("contractexplorer/[controller]")]
+//    public class ContractExplorerController : Controller
+//    {
+
+//    }
+//}
