+++ src/Stratis.Bitcoin.IntegrationTests/Mempool/MemoryPoolTests.cs
@@ -8,7 +8,6 @@ using Stratis.Bitcoin.Features.MemoryPool;
 using Stratis.Bitcoin.IntegrationTests.Common;
 using Stratis.Bitcoin.IntegrationTests.Common.EnvironmentMockUpHelpers;
 using Stratis.Bitcoin.Tests.Common;
-using Stratis.Bitcoin.Utilities;
 using Xunit;
 
 namespace Stratis.Bitcoin.IntegrationTests.Mempool
@@ -20,22 +19,6 @@ namespace Stratis.Bitcoin.IntegrationTests.Mempool
         public MemoryPoolTests()
         {
             this.network = KnownNetworks.RegTest;
-        }
-
-        public class DateTimeProviderSet : DateTimeProvider
-        {
-            public long time;
-            public DateTime timeutc;
-
-            public override long GetTime()
-            {
-                return this.time;
-            }
-
-            public override DateTime GetUtcNow()
-            {
-                return this.timeutc;
-            }
         }
 
         [Fact]
@@ -399,55 +382,6 @@ namespace Stratis.Bitcoin.IntegrationTests.Mempool
                 TestHelper.WaitLoop(() => stratisNode1.CreateRPCClient().GetRawMempool().Length == 0);
                 TestHelper.WaitLoop(() => stratisNode2.CreateRPCClient().GetRawMempool().Length == 0);
             }
-        }
-    }
-
-    public class TestMemPoolEntryHelper
-    {
-        // Default values
-        private Money nFee = Money.Zero;
-
-        private long nTime = 0;
-        private double dPriority = 0.0;
-        private int nHeight = 1;
-        private bool spendsCoinbase = false;
-        private long sigOpCost = 4;
-        private LockPoints lp = new LockPoints();
-
-        public TxMempoolEntry FromTx(Transaction tx, TxMempool pool = null)
-        {
-            Money inChainValue = (pool != null && pool.HasNoInputsOf(tx)) ? tx.TotalOut : 0;
-
-            return new TxMempoolEntry(tx, this.nFee, this.nTime, this.dPriority, this.nHeight,
-                inChainValue, this.spendsCoinbase, this.sigOpCost, this.lp, new ConsensusOptions());
-        }
-
-        // Change the default value
-        public TestMemPoolEntryHelper Fee(Money fee) { this.nFee = fee; return this; }
-
-        public TestMemPoolEntryHelper Time(long time)
-        {
-            this.nTime = time; return this;
-        }
-
-        public TestMemPoolEntryHelper Priority(double priority)
-        {
-            this.dPriority = priority; return this;
-        }
-
-        public TestMemPoolEntryHelper Height(int height)
-        {
-            this.nHeight = height; return this;
-        }
-
-        public TestMemPoolEntryHelper SpendsCoinbase(bool flag)
-        {
-            this.spendsCoinbase = flag; return this;
-        }
-
-        public TestMemPoolEntryHelper SigOpsCost(long sigopsCost)
-        {
-            this.sigOpCost = sigopsCost; return this;
         }
     }
 }
