+++ src/Stratis.SmartContracts.IntegrationTests/ContractCreationTests.cs
@@ -10,7 +10,7 @@ using Stratis.SmartContracts.Executor.Reflection;
 using Stratis.SmartContracts.Executor.Reflection.Compilation;
 using Xunit;
 
-namespace Stratis.Bitcoin.IntegrationTests.SmartContracts
+namespace Stratis.SmartContracts.IntegrationTests
 {
     public class ContractCreationTests
     {
