+++ src/Stratis.SmartContracts.IntegrationTests/MockChain/MockChain.cs
