+++ src/Stratis.SmartContracts.IntegrationTests/MockChain/MockChainNode.cs
