+++ src/Stratis.SmartContracts.IntegrationTests/SmartContracts/Auction.cs
