+++ src/Stratis.SmartContracts.IntegrationTests/SmartContracts/CallContract.cs
