+++ src/Stratis.SmartContracts.IntegrationTests/SmartContracts/ContractCreation.cs
