+++ src/Stratis.SmartContracts.IntegrationTests/SmartContracts/CountContract.cs
