+++ src/Stratis.SmartContracts.IntegrationTests/SmartContracts/InterContract1.cs
