+++ src/Stratis.SmartContracts.IntegrationTests/SmartContracts/InterContract2.cs
