+++ src/Stratis.SmartContracts.IntegrationTests/SmartContracts/ReceiveHandlerContract.cs
