+++ src/Stratis.SmartContracts.IntegrationTests/SmartContracts/StorageDemo.cs
