+++ src/Stratis.SmartContracts.IntegrationTests/SmartContracts/Token.cs
