+++ src/Stratis.SmartContracts.IntegrationTests/SmartContracts/TransferTestPos.cs
